// definitions/RpgLogs.d.ts
export namespace RpgLogs {
// ...
    type ActorType = 'Player' | 'Pet' | 'NPC';

    export interface Actor {
        /**
         * The id of the actor.
         */
        id: number;

        /**
         * The id of the actor within its containing report.
         */
        idInReport: number;

        /**
         * The game id of the actor.
         */
        gameId: number;

        /**
         * The name of the actor in the user's desired language.
         */
        name: string;

        /**
         * The type of the actor. This will be Player, Pet or NPC.
         */
        type: ActorType;

        /**
         * The subType of the actor. This will be the job (for FF) or class (for other games) of players. NPCs can be "Boss" or "NPC".
         */
        subType: string;
    }
// ...
    export interface ResourceData {
        /**
         * The unit's current hit points.
         */
        hitPoints: number;

        /**
         * The unit's maximum hit points.
         */
        maxHitPoints: number;

        /**
         * The unit's x position.
         */
        x: number;

        /**
         * The unit's y position.
         */
        y: number;

        /**
         * The unit's facing. Null for games that do not support facing.
         */
        facing: number | null;
    }
// ...
    export interface Event {
        /**
         * The time offset of the event in milliseconds from the start of the report.
         */
        timestamp: number;

        /**
         * The source actor. Null if no source is set.
         */
        source: Actor | null;

        /**
         * The instance id of the source.
         */
        sourceInstanceId: number;

        /**
         * The raw disposition of the source. Use the isFriendlyParticipant and isEnemyParticipant methods to obtain
         * the overall single disposition for this actor. This field should be used with caution as it could flip if the
         * logger or the unit get mind controlled.
         */
        sourceDisposition: ActorDisposition;

        /**
         * The raid marker on this source unit. 0 if no raid marker is set.
         */
        sourceRaidMarker: number;

        /**
         * Resource data for the source. Null if no resource information is available.
         */
        sourceResources: ResourceData | null;

        /**
         * The target actor. Null if no target is set.
         */
        target: Actor | null;

        /**
         * The instance id of the target.
         */
        targetInstanceId: number;

        /**
         * The raw disposition of the target. Use the isFriendlyParticipant and isEnemyParticipant methods on Fight to obtain
         * the overall single disposition for this actor. This field should be used with caution as it could flip if the
         * logger or the unit get mind controlled.
         */
        targetDisposition: ActorDisposition;

        /**
         * The raid marker on this source unit. 0 if no raid marker is set.
         */
        targetRaidMarker: number;

        /**
         * Resource data for the target. Null if no resource information is available.
         */
        targetResources: ResourceData | null;

        /**
         * The ability used by the source on the target. Null if no ability is set.
         */
        ability: Ability | null;
    }
// ...
    export interface CastEvent extends Event {
        /**
         * The type of the event.
         */
        type: 'cast';

        /**
         * Whether or not the cast is a melee swing/auto-attack.
         */
        isMelee: boolean;

        /**
         * Whether or not the cast is fake, i.e., not the result of a user action, but just made up by the game.
         */
        isFake: boolean;
    }
// ...
    export type TableComponent = {
        component: 'Table';
        props: {
            columns: Record<string, TableColumn>;
            data: readonly Record<string, unknown>[];
        };
    };

    type TableColumn = {
        header: string;
        width?: number;
        maxWidth?: number;
        minWidth?: number;
        noWrap?: boolean;
        textAlign?: 'left' | 'center' | 'right';
        columns?: Record<string, TableColumn>;
    };
}

export namespace Highcharts {
    type NYI = unknown;
// ...
    type Color = string | GradientColorObject;
// ...
    namespace Series {
// ...
        export type Any = KnownSeries | UnknownSeries;
    }

    type AsOptions<T> = Omit<T, 'data'>;

    type TextOptions = Partial<{
        align: 'left' | 'center' | 'right';
        floating: boolean;
        margin: number;
        /**
         * Text to display.
         *
         * Note that while Highcharts supports HTML, we do not. HTML-like tokens will be escaped.
         * This will eventually be converted to Markdown, but that is NYI.
         *
         * This note applies to all fields labeled with the `Html` type.
         */
        text: Html;
        verticalAlign: 'top' | 'middle' | 'bottom';
        x: number;
        y: number;
        useHTML: boolean;
        style: Css;
    }>;

    type TooltipOptions = Partial<{
        animation: boolean;
        backgroundColor: Color;
        borderColor: Color;
        borderRadius: number;
        borderWidth: number;
        className: Unsupported;
        dateTimeLabelFormats: NYI;
        distance: number;
        enabled: boolean;
        followPointer: boolean;
        followTouchMove: boolean;
        clusterFormat: Html;
        footerFormat: Html;
        formatter: Unsupported;
        headerFormat: Html;
        headerShape: NYI;
        hideDelay: number;
        nullFormat: Html;
        nullFormatter: Unsupported;
        outside: boolean;
        padding: number;
        pointFormat: Html;
        positioner: Unsupported;
        shadow: boolean | NYI;
        shape: 'callout' | 'circle' | 'square';
        shared: boolean;
        snap: number;
        split: boolean;
        stickOnContact: boolean;
        style: Css;
        useHTML: boolean;
        valueDecimals: number | undefined;
        valuePrefix: string;
        valueSuffix: string;
        xDateFormat: string;
    }>;

    type AxisOptions = Partial<{
        accessibility: NYI;
        alignTicks: boolean;
        allowDecimals: boolean;
        alternateGridColor: Color;
        angle: number;
        breaks: NYI;
        categories: string[];
        ceiling: number;
        className: Unsupported;
        crosshair: NYI;
        dateTimeLabelFormats: NYI;
        endOnTick: boolean;
        events: Unsupported;
        floor: number;
        gridLineColor: Color;
        gridLineDashStyle: DashStyle;
        gridLineInterpolation: 'circle' | 'polygon';
        gridLineWidth: number;
        gridZIndex: number;
        height: number | string;
        id: string;
        labels: NYI;
        left: number | string;
        lineColor: Color;
        lineWidth: number;
        linkedTo: number;
        margin: number;
        max: number | null;
        maxPadding: number;
        min: number | null;
        minorGridLineColor: Color;
        minorGridLineDashStyle: DashStyle;
        minorGridLineWidth: number;
        minorTickColor: Color;
        minorTickInterval: number | string | null;
        minorTickLength: number;
        minorTickPosition: string;
        minorTicks: boolean;
        minorTickWidth: number;
        minPadding: number;
        minRange: number;
        minTickInterval: number;
        offset: number;
        opposite: boolean;
        pane: number;
        panningEnabled: boolean;
        plotBands: NYI;
        plotLines: NYI;
        reversed: boolean;
        reversedStacks: boolean;
        showEmpty: boolean;
        showFirstLabel: boolean;
        showLastLabel: boolean;
        softMax: number;
        softMin: number;
        startOfWeek: number;
        startOnTick: boolean;
        tickAmount: number;
        tickColor: Color;
        tickInterval: number;
        tickLength: number;
        tickmarkPlacement: string;
        tickPixelInterval: number;
        tickPosition: string;
        tickPositioner: Unsupported;
        tickPositions: number[];
        tickWidth: number | undefined;
        title: NYI;
        top: number | string;
        type: 'linear' | 'logarithmic' | 'datetime' | 'category';
        uniqueNames: boolean;
        units: NYI;
        visible: boolean;
        width: number | string;
        zIndex: number;
        zoomEnabled: boolean;
    }>;

    type ChartOptions = NYI;

    export type Options = {
        plotOptions?: {
            [Key in Series.Any['type']]: AsOptions<
                Extract<Series.Any, { type: Key }>
                >;
        };
        colors?: Color[];
        series: Series.Any[];
        caption?: TextOptions;
        title?: TextOptions;
        subtitle?: TextOptions;
        tooltip?: TooltipOptions;
        xAxis?: AxisOptions;
        yAxis?: AxisOptions;
        zAxis?: AxisOptions;
        colorAxis?: NYI; // does not really share axis options with the x/y/z axes
        chart?: ChartOptions;
        annotations?: NYI[];
        legend?: NYI;
        labels?: NYI;
        pane?: NYI;
        time?: NYI;
    };
}

export namespace RpgLogs {
    // this lint is disabled because interfaces show the new name in the Monaco editor, but type aliases do not
    // eslint-disable-next-line @typescript-eslint/no-empty-interface
    export interface ChartComponentProps extends Highcharts.Options {}
}